requirements.txt
paramiko
docker
pyyaml
click
sqlalchemy
flake8
